requirements.txt
numpy>=2.4.0,<3.0
scipy>=1.17.0,<2.0
matplotlib>=3.10.0,<4.0
plotly>=6.6.0,<7.0
pyyaml>=6.0,<7.0
h5py>=3.16.0,<4.0